pyproject.toml
[project]
name = "poly-scout"
version = "0.1.0"
description = "Polymarket wallet scanner - detect emerging alpha traders"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "asyncio>=3.4.3",
    "rich>=13.7.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.2.1",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
poly-scout = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
